+++ app/src/main/java/com/example/kyc_kotlin_sample/SplashActivity.kt
@@ -1,31 +1,24 @@
-package com.example.kyc_kotlin_sample;
+package com.example.kyc_kotlin_sample
 
-import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.content.Intent
+import android.os.Handler
 
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-public class SplashActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash);
-
-        Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent = new Intent(getApplicationContext(),MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        },3000);
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+        val handler = Handler()
+        handler.postDelayed({
+            val intent = Intent(applicationContext, MainActivity::class.java)
+            startActivity(intent)
+            finish()
+        }, 3000)
     }
-    @Override
-    protected void onPause() {
-        super.onPause();
-        finish();
+
+    override fun onPause() {
+        super.onPause()
+        finish()
     }
 }
