--- app/src/main/java/com/example/kyc_kotlin_sample/WebViewActivity.java
@@ -1,4 +1,0 @@
-package com.example.kyc_kotlin_sample;
-
-public class WebViewActivity {
-}
